+++ test/test_multithread.cc
@@ -34,7 +34,7 @@ OperationWindow GenMTSmallWindow() {
   window.records.reserve(length);
   std::default_random_engine gen;
   std::uniform_int_distribution<size_t> dis(window.left / 8, window.right / 8);
-  for (int i = 0; i < length; ++i) {
+  for (size_t i = 0; i < length; ++i) {
     window.records.push_back({nullptr, dis(gen) * 8, 0});
   }
 
@@ -55,14 +55,14 @@ void ThreadTest(barrier *barr) {
   std::cout << oss.str() << std::flush;
 
   for (size_t t = 0; t < window.times; ++t) {
-    for (int i = 0; i < window.records.size(); ++i) {
+    for (size_t i = 0; i < window.records.size(); ++i) {
       window.records[i].ptr = static_cast<uint64_t*>(
           Static::thread_alloc()->ThreadAlloc(window.records[i].mem_size)
       );
       // std::cout << "alloc: " << window.records[i].ptr << std::endl;
       ASSERT_NE(nullptr, window.records[i].ptr);
     }
-    for (int i = 0; i < window.records.size(); ++i) {
+    for (size_t i = 0; i < window.records.size(); ++i) {
       // std::cout << "free : " << window.records[i].ptr << std::endl;
       void *ptr = window.records[i].ptr;
       Static::thread_alloc()->ThreadDalloc(ptr);
@@ -71,12 +71,12 @@ void ThreadTest(barrier *barr) {
 }
 
 TEST(MultiThread, C1T1) {
-  unsigned int num_threads = 1;
+  size_t num_threads = 1;
   Static::set_num_arena(1);
   barrier barr(num_threads);
 
   std::vector<thread> threads;
-  for (unsigned int i = 0; i < num_threads; ++i) {
+  for (size_t i = 0; i < num_threads; ++i) {
     threads.push_back(thread(ThreadTest, &barr));
   }
   for (size_t i = 0; i < threads.size(); ++i) {
@@ -86,12 +86,12 @@ TEST(MultiThread, C1T1) {
 
 TEST(MultiThread, C8T8) {
   // TODO
-  unsigned int num_threads = 8;
+  size_t num_threads = 8;
   Static::set_num_arena(8);
   barrier barr(num_threads);
 
   std::vector<thread> threads;
-  for (unsigned int i = 0; i < num_threads; ++i) {
+  for (size_t i = 0; i < num_threads; ++i) {
     threads.push_back(thread(ThreadTest, &barr));
   }
   for (size_t i = 0; i < threads.size(); ++i) {
@@ -101,12 +101,12 @@ TEST(MultiThread, C8T8) {
 
 TEST(MultiThread, C1T8) {
   // TODO
-  unsigned int num_threads = 8;
+  size_t num_threads = 8;
   Static::set_num_arena(1);
   barrier barr(num_threads);
 
   std::vector<thread> threads;
-  for (unsigned int i = 0; i < num_threads; ++i) {
+  for (size_t i = 0; i < num_threads; ++i) {
     threads.push_back(thread(ThreadTest, &barr));
   }
   for (size_t i = 0; i < threads.size(); ++i) {
@@ -116,12 +116,12 @@ TEST(MultiThread, C1T8) {
 
 TEST(MultiThread, C8T64) {
   // TODO
-  unsigned int num_threads = 64;
+  size_t num_threads = 64;
   Static::set_num_arena(8);
   barrier barr(num_threads);
 
   std::vector<thread> threads;
-  for (unsigned int i = 0; i < num_threads; ++i) {
+  for (size_t i = 0; i < num_threads; ++i) {
     threads.push_back(thread(ThreadTest, &barr));
   }
   for (size_t i = 0; i < threads.size(); ++i) {
